core: events: bpf: do not copy events twice (raw and final event)
Refine the logic of parse_raw_section & parse_single_raw_section to
retrieve a typed reference to the raw event stored in a slice. This way
we don't copy the event to the raw representation before copying it
again to the final event one.

It's fine to do this in parse_single_raw_section especially as we know
for sure we don't have more than one sections, so we don't need to pop
them.

Improves the performance of parse_raw_event by 3%.

Signed-off-by: Antoine Tenart <atenart@redhat.com>