core: probe: bpf: common: fill the generic event part only if needed
The generic event part of the event was filed regardless of if hooks
added their own sections. In case they did not, the common parts were
discarded. Only fill the generic part if hook added data in the event,
to optimize for cases when they do not.

A few extra changes to keep the verifier happy.

Signed-off-by: Antoine Tenart <atenart@redhat.com>